% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aliquot.R
\docType{class}
\name{Aliquot}
\alias{Aliquot}
\title{Aliquot class}
\format{\code{\link{Aliquot}} class generator}
\usage{
\code{al = Aliquot$new()}
}
\description{
A class representation of a cytokine aliquot.
}
\keyword{data}

