% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guru-utils.R
\name{alpha_to_guru}
\alias{alpha_to_guru}
\title{Compute the LabGuru box position}
\usage{
alpha_to_guru(loc_string, reverse = FALSE, box_cols = 9)
}
\arguments{
\item{loc_string}{The value to convert}

\item{reverse}{Which direction do you want to go}

\item{box_cols}{How many columns are in the storage box}
}
\description{
This function takes an alphanumeric box location (ex. B2) and computes the
corresponding numerical position used in LabGuru. Passing
\code{reverse = TRUE} allows users to perform the computation in revers --
i.e. to compute the alphanumeric position from the LabGuru position.
}

