% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/guru-utils.R
\name{check_guru_key}
\alias{check_guru_key}
\title{Check if an existing API key works}
\usage{
check_guru_key(token)
}
\arguments{
\item{token}{The API token to test}
}
\description{
This function will test an API token from LabGuru - returning \code{TRUE} if
the token is valid and \code{FALSE} otherwise.
}

