% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/redcap-utils.R
\name{create_api_config}
\alias{create_api_config}
\title{Create API configuration}
\usage{
create_api_config(api_url = NULL, token = NULL,
  config_options = httr::config(cainfo =
  system.file("ssl_certs/mozilla_ca_root.crt", package = "REDCapR"), timeout =
  8, ssl_verifypeer = 1))
}
\arguments{
\item{api_url}{The API URL of the REDCap instance you'd like to connect to.}

\item{token}{The API token for the project you'd like to connect to.}

\item{config_options}{A list of \code{CURL} options to be passed to
 \code{httr::POST}. See ?httr::config for full details.}
}
\description{
Function is used to create a named list of API configuration to be
passed to any of the methods that interact with the
REDCap API. Current defaults include using SSL certificates from
\code{REDCapR} package.
}

