% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/guru-utils.R
\name{create_box}
\alias{create_box}
\title{Create a box in LabGuru}
\usage{
create_box(box_name, token, box_rows = 9, box_cols = 9, type = "Tissue",
  color = "gray")
}
\arguments{
\item{box_name}{The name of the box to create}

\item{token}{A valid LabGuru API token}

\item{box_rows}{Number of rows in the box}

\item{box_cols}{Number of columns in the box}

\item{data_type}{What kind of object tube, tissue, etc}
}
\description{
This function lets you create a box via the LabGuru API.
}

