% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/guru-utils.R
\name{create_tissue}
\alias{create_tissue}
\title{Create a tissue in LabGuru}
\usage{
create_tissue(base_id, descr, token, type = "blood")
}
\description{
This function lets you create a box via the LabGuru API.
}

