% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guru-utils.R
\name{create_tube}
\alias{create_tube}
\title{Create a tube in LabGuru}
\usage{
create_tube(tube_name, tube_barcode, box, box_location, tissue_uuid, tube_notes,
  token)
}
\description{
This function lets you create a tube via the LabGuru API.
}

