% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guru-utils.R
\name{get_all}
\alias{get_all}
\title{Get all of something from LabGuru}
\usage{
get_all(token, data_type = "boxes")
}
\arguments{
\item{token}{A valid LabGuru API token}

\item{data_type}{What kind of object tube, tissue, etc}
}
\description{
This function allows you to pull all of something (tube, tissue,
etc) from the LabGuru API.
}
\examples{
\dontrun{
# get a box via the api
tok <- 'API_SECRET'
box_info <- get_all(token = tok, data_type = 'boxes')

}
}

