% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/redcap-utils.R
\name{get_data_dict}
\alias{get_data_dict}
\title{Get data dictionary as \code{data.frame}}
\usage{
get_data_dict(api_config)
}
\arguments{
\item{api_config}{A list resulting from \code{create_api_config}}
}
\description{
\code{get_data_dict} returns a \code{data.frame} containing the project's Data Dictionary
}

