% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guru-utils.R
\name{get_guru_key}
\alias{get_guru_key}
\title{Get a new API key}
\usage{
get_guru_key(user_name, password)
}
\arguments{
\item{user_name}{Your LabGuru user name}

\item{password}{Your LabGuru password}
}
\description{
This function will get an updated API token from LabGuru to replace an
expired one.
}
\examples{
\dontrun{
# get a new token
my_secrets = readRDS('~/.labguru.rds')
new_api_key <- get_guru_key(my_secrets[['user']], my_secrets[['pass']])
}

}

