% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guru-utils.R
\name{get_one}
\alias{get_one}
\title{Get one of something from LabGuru}
\usage{
get_one(token, id, data_type)
}
\arguments{
\item{token}{A valid LabGuru API token}

\item{id}{The ID of the object you would like to retrieve}

\item{data_type}{What kind of object \code{tubes}, \code{tissues}, or \code{boxes}}
}
\description{
This function allows you to pull something (tube, tissue, etc) from
the LabGuru API.
}
\examples{
\dontrun{
# get a box via the api
tok <- 'API_SECRET'
box_info <- get_one(token = tok, id = 70, data_type = 'boxes')

}
}

