% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/redcap-utils.R
\name{get_redcap_values}
\alias{get_redcap_values}
\title{Get some values out of REDCap}
\usage{
get_redcap_values(api_payload, fnames, record_id)
}
\arguments{
\item{fnames}{A vector of field names in REDCap}

\item{record_id}{A REDCap ID for a record in the project}

\item{api_config}{A list resulting from \code{create_api_config}}
}
\description{
Returns a data frame with values for the specified field names for the
specified patient
}

