% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Patient.R
\docType{class}
\name{Patient}
\alias{Patient}
\title{Patient class}
\format{\code{\link{Patient}} class generator}
\usage{
\code{pat = Patient$new()}
}
\description{
A class representation of a patient.
}
\keyword{data}

