% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plate.R
\docType{class}
\name{Plate}
\alias{Plate}
\title{Plate class}
\format{\code{\link{Plate}} class generator}
\usage{
\code{plate = Plate$new()}
}
\description{
A class representation of a Bio-Rad plate.
}
\keyword{data}

