% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SyncHandler.R
\docType{class}
\name{SyncHandler}
\alias{SyncHandler}
\title{SyncHandler class}
\format{\code{\link{SyncHandler}} class generator}
\usage{
\code{sync = SyncHandler$new()}
}
\description{
A class to handle syncing our data in REDCap, DB, and LabGuru.
}
\keyword{data}

