% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/guru-utils.R
\name{update_tissue}
\alias{update_tissue}
\title{Update a tissue in LabGuru}
\usage{
update_tissue(id, name, token, description = NULL, tissue_type = NULL)
}
\arguments{
\item{id}{The numeric ID of the tissue}

\item{name}{The character name of the tissue}

\item{token}{A valid API token}

\item{description}{An updated description of the tissue}

\item{tissue_type}{An updated description of the tissue}
}
\description{
This function lets you update an existing tissue via the LabGuru API.
}

