% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyncHandler.R
\docType{class}
\name{SyncHandler}
\alias{SyncHandler}
\title{SyncHandler class}
\format{\code{\link{SyncHandler}} class generator}
\usage{
\code{
sync <- SyncHandler$new(project_id, instrument, redcap_id, redcap_token,
                       guru_token, redcap_url)
sync$sync_data()
}
}
\description{
A class to handle syncing our data in REDCap, DB, and LabGuru.
}
\keyword{data}

